only update new token or new os is arrived

+++ Covid/src/routers/user.ts
@@ -15,8 +15,10 @@ router.post('/sign', async (req, res) => {
         if (!token || !uuid || !os) throw new Error('No Appropriate fields in Request')
         const found = await userDao.getUserByUUID(uuid)
         if (found) {
-            await userDao.deleteUserByUUID(uuid)
-            await userDao.insertUser(token, uuid, os)
+            if (found.token !== token || found.os !== os) {
+                await userDao.deleteUserByUUID(uuid)
+                await userDao.insertUser(token, uuid, os)
+            }
             sendRes(res, 201, '이미 존재하는 유저입니다', {
                 updated: true
             })
